Adding files and updates
Here I have completed the updates of more files, as well as updated. The goal is to have all the files have a good introduction, background, usage, and verify the files and modules do what I want them to do functionally. I'm also updating to modern convention of using the main() function definitions